refactor: util 함수에 대한 type은 type 폴더로 분리
