+++ src/domain/utils/types/RandomNumber.ts
